Before, we used sglang frontend which does not deliver high throughput, the [new benchmark of sglang](https://lmsys.org/blog/2024-07-25-sglang-llama3/) is based on sglang runtime, thus in this PR we switched to srt.

# Benchmark

## Before

```
Total time: 38.64 s
Throughput: 0.78 requests/s
Average latency: 38.63 s
Average latency per token: 0.17 s
Average latency per output token: 1.11 s
Throughput: 145.40378699622423 tokens/s
```

## After

```
Total time: 6.02 s
Throughput: 4.98 requests/s
Average latency: 2.44 s
Average latency per token: 0.01 s
Average latency per output token: 0.03 s
Throughput: 959.2033185347453 tokens/s
```